+++ amnesia/src/game/LuxConfigHandler.cpp
@@ -85,6 +85,7 @@ void cLuxConfigHandler::LoadMainConfig()
 	mbUncapFPS = gpBase->mpMainConfig->GetBool("Engine","LimitFPS", false) == false;
 	mlSimulationRate = gpBase->mpMainConfig->GetInt("Engine","SimulationRate", 60);
 	if(mlSimulationRate < 60) mlSimulationRate = 60;
+	mbSleepWhenOutOfFocus = gpBase->mpMainConfig->GetBool("Engine","SleepWhenOutOfFocus", true);
 
 	mbFastPhysicsLoad=	gpBase->mpMainConfig->GetBool("MapLoad","FastPhysicsLoad", false);
 	mbFastStaticLoad=	gpBase->mpMainConfig->GetBool("MapLoad","FastStaticLoad", false);
@@ -204,7 +205,7 @@ void cLuxConfigHandler::SaveMainConfig()
 	// Engine properties
 	gpBase->mpMainConfig->SetBool("Engine","LimitFPS", mbUncapFPS==false);
 	gpBase->mpMainConfig->SetInt("Engine","SimulationRate", mlSimulationRate);
-	gpBase->mpMainConfig->SetBool("Engine","SleepWhenOutOfFocus",gpBase->mpEngine->GetWaitIfAppOutOfFocus());
+	gpBase->mpMainConfig->SetBool("Engine","SleepWhenOutOfFocus", mbSleepWhenOutOfFocus);
 }
 
 //-----------------------------------------------------------------------